test(dbt): move tests into `core/` and `legacy/`
