+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/conftest.py
@@ -1,11 +1,9 @@
 import os
-import subprocess
 from pathlib import Path
 from typing import Any, Dict
 
 import pytest
-from dagster._utils import file_relative_path, pushd
-from dagster_dbt import DbtCliClientResource, DbtCliResource, dbt_cli_resource
+from dagster_dbt import DbtCliResource
 from dagster_dbt.core.resources_v2 import DbtCliInvocation
 
 from .dbt_projects import (
@@ -18,83 +16,6 @@ from .dbt_projects import (
     test_meta_config_path,
     test_metadata_path,
 )
-
-# ======= CONFIG ========
-DBT_EXECUTABLE = "dbt"
-TEST_PROJECT_DIR = file_relative_path(__file__, "dagster_dbt_test_project")
-DBT_CONFIG_DIR = TEST_PROJECT_DIR
-TEST_DBT_TARGET_DIR = os.path.join(TEST_PROJECT_DIR, "target_test")
-
-TEST_PYTHON_PROJECT_DIR = file_relative_path(__file__, "dagster_dbt_python_test_project")
-DBT_PYTHON_CONFIG_DIR = TEST_PYTHON_PROJECT_DIR
-
-IS_BUILDKITE = os.getenv("BUILDKITE") is not None
-
-
-@pytest.fixture(scope="session")
-def test_project_dir():
-    return TEST_PROJECT_DIR
-
-
-@pytest.fixture(scope="session")
-def dbt_config_dir():
-    return DBT_CONFIG_DIR
-
-
-@pytest.fixture(scope="session")
-def dbt_target_dir():
-    return TEST_DBT_TARGET_DIR
-
-
-@pytest.fixture(scope="session")
-def dbt_executable():
-    return DBT_EXECUTABLE
-
-
-@pytest.fixture(scope="session")
-def test_python_project_dir():
-    return TEST_PYTHON_PROJECT_DIR
-
-
-@pytest.fixture(scope="session")
-def dbt_python_config_dir():
-    return DBT_PYTHON_CONFIG_DIR
-
-
-@pytest.fixture(
-    scope="session",
-    params=[
-        pytest.param("legacy", marks=pytest.mark.legacy),
-        pytest.param("DbtCliClientResource", marks=pytest.mark.legacy),
-        pytest.param("DbtCliResource"),
-    ],
-)
-def dbt_cli_resource_factory(request):
-    if request.param == "DbtCliClientResource":
-        return lambda **kwargs: DbtCliClientResource(
-            project_dir=kwargs["project_dir"],
-            profiles_dir=kwargs.get("profiles_dir"),
-            json_log_format=kwargs.get("json_log_format", True),
-        )
-    elif request.param == "DbtCliResource":
-        return lambda **kwargs: DbtCliResource(
-            project_dir=kwargs["project_dir"], profile=kwargs.get("profile")
-        )
-    else:
-        return lambda **kwargs: dbt_cli_resource.configured(kwargs)
-
-
-@pytest.fixture(scope="session")
-def dbt_seed(dbt_executable, dbt_config_dir):
-    with pushd(TEST_PROJECT_DIR):
-        subprocess.run([dbt_executable, "seed", "--profiles-dir", dbt_config_dir], check=True)
-
-
-@pytest.fixture(scope="session")
-def dbt_build(dbt_executable, dbt_config_dir):
-    with pushd(TEST_PROJECT_DIR):
-        subprocess.run([dbt_executable, "seed", "--profiles-dir", dbt_config_dir], check=True)
-        subprocess.run([dbt_executable, "run", "--profiles-dir", dbt_config_dir], check=True)
 
 
 def _create_dbt_invocation(project_dir: Path) -> DbtCliInvocation:
