+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/legacy/dagster_dbt_python_test_project/models/daily_aggregated_events.sql
