+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/legacy/dagster_dbt_test_project/macros/log_macro.sql
