+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/legacy/dagster_dbt_test_project/models/sort_cold_cereals_by_calories.sql
